requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pytest==8.3.2
httpx==0.27.2
python-multipart==0.0.9
pywinrm==0.4.3
